token/backend/requirements.txt
Flask
flask-cors
jinja2
waitress
psycopg2-binary
